Reorganize data structures under own modules
