+++ bokehjs/src/lib/models/glyphs/glyph.ts
@@ -8,7 +8,8 @@ import {View} from "core/view"
 import {Model} from "../../model"
 import {Anchor} from "core/enums"
 import {logger} from "core/logging"
-import {Arrayable, Rect, NumberArray, RaggedArray, Indices} from "core/types"
+import {Arrayable, Rect, NumberArray, Indices} from "core/types"
+import {RaggedArray} from "core/util/ragged_array"
 import {map, max} from "core/util/arrayable"
 import {SpatialIndex} from "core/util/spatial"
 import {Scale} from "../scales/scale"
